+++ live-demo/serve.go
@@ -13,7 +13,7 @@ import (
 var G goose.Alert = goose.Alert(2)
 
 func main() {
-	fs := http.FileServer(http.Dir("docs"))
+	fs := http.FileServer(http.Dir("../docs"))
 	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
 		if strings.HasSuffix(r.URL.Path, ".wasm") {
 			w.Header().Set("Content-Type", "application/wasm")